feat: 유저 상태에 따른 PrivateRoute 적용 (#26) - router 파일 위치 수정
